reuse leaderboard for SoloResultsScreen
