+++ osu.Game/Screens/OnlinePlay/Multiplayer/Spectate/MultiSpectatorResultsScreen.cs
@@ -3,9 +3,6 @@
 
 #nullable disable
 
-using System;
-using System.Collections.Generic;
-using osu.Game.Online.API;
 using osu.Game.Scoring;
 using osu.Game.Screens.Play;
 
@@ -24,9 +21,5 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer.Spectate
 
             Scheduler.AddDelayed(() => StatisticsPanel.ToggleVisibility(), 1000);
         }
-
-        protected override APIRequest FetchScores(Action<IEnumerable<ScoreInfo>> scoresCallback) => null;
-
-        protected override APIRequest FetchNextPage(int direction, Action<IEnumerable<ScoreInfo>> scoresCallback) => null;
     }
 }
