+++ osu.Game/Screens/Play/Player.cs
@@ -1224,11 +1224,7 @@ namespace osu.Game.Screens.Play
         /// </summary>
         /// <param name="score">The <see cref="ScoreInfo"/> to be displayed in the results screen.</param>
         /// <returns>The <see cref="ResultsScreen"/>.</returns>
-        protected virtual ResultsScreen CreateResults(ScoreInfo score) => new SoloResultsScreen(score)
-        {
-            AllowRetry = true,
-            ShowUserStatistics = true,
-        };
+        protected abstract ResultsScreen CreateResults(ScoreInfo score);
 
         private void fadeOut(bool instant = false)
         {
