+++ osu.Game/Screens/Play/ReplayPlayer.cs
@@ -92,7 +92,10 @@ namespace osu.Game.Screens.Play
                 Scores = { BindTarget = LeaderboardScores }
             };
 
-        protected override ResultsScreen CreateResults(ScoreInfo score) => new SoloResultsScreen(score);
+        protected override ResultsScreen CreateResults(ScoreInfo score) => new SoloResultsScreen(score)
+        {
+            Scores = { BindTarget = LeaderboardScores }
+        };
 
         public bool OnPressed(KeyBindingPressEvent<GlobalAction> e)
         {
