+++ osu.Game/Screens/Play/SoloPlayer.cs
@@ -14,6 +14,7 @@ using osu.Game.Online.Rooms;
 using osu.Game.Online.Solo;
 using osu.Game.Scoring;
 using osu.Game.Screens.Play.HUD;
+using osu.Game.Screens.Ranking;
 
 namespace osu.Game.Screens.Play
 {
@@ -72,5 +73,12 @@ namespace osu.Game.Screens.Play
 
             return new SubmitSoloScoreRequest(score.ScoreInfo, token, beatmap.OnlineID);
         }
+
+        protected override ResultsScreen CreateResults(ScoreInfo score) => new SoloResultsScreen(score)
+        {
+            AllowRetry = true,
+            ShowUserStatistics = true,
+            Scores = { BindTarget = LeaderboardScores }
+        };
     }
 }
