+++ osu.Game/Screens/Ranking/SoloResultsScreen.cs
@@ -1,49 +1,40 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using osu.Framework.Allocation;
-using osu.Game.Beatmaps;
-using osu.Game.Extensions;
-using osu.Game.Online.API;
-using osu.Game.Online.API.Requests;
-using osu.Game.Rulesets;
+using osu.Framework.Bindables;
 using osu.Game.Scoring;
 
 namespace osu.Game.Screens.Ranking
 {
     public partial class SoloResultsScreen : ResultsScreen
     {
-        private GetScoresRequest? getScoreRequest;
+        public readonly IBindableList<ScoreInfo> Scores = new BindableList<ScoreInfo>();
 
-        [Resolved]
-        private RulesetStore rulesets { get; set; } = null!;
+        private bool hasLoaded;
 
         public SoloResultsScreen(ScoreInfo score)
             : base(score)
         {
         }
 
-        protected override APIRequest? FetchScores(Action<IEnumerable<ScoreInfo>>? scoresCallback)
+        protected override void PopulateScorePanelList()
         {
-            Debug.Assert(Score != null);
-
-            if (Score.BeatmapInfo!.OnlineID <= 0 || Score.BeatmapInfo.Status <= BeatmapOnlineStatus.Pending)
-                return null;
-
-            getScoreRequest = new GetScoresRequest(Score.BeatmapInfo, Score.Ruleset);
-            getScoreRequest.Success += r => scoresCallback?.Invoke(r.Scores.Where(s => !s.MatchesOnlineID(Score)).Select(s => s.ToScoreInfo(rulesets, Beatmap.Value.BeatmapInfo)));
-            return getScoreRequest;
+            Scores.BindCollectionChanged((_, _) => Scheduler.AddOnce(addScores), true);
         }
 
-        protected override void Dispose(bool isDisposing)
+        private void addScores()
         {
-            base.Dispose(isDisposing);
+            if (!Scores.Any() || hasLoaded)
+                return;
 
-            getScoreRequest?.Cancel();
+            foreach (var s in Scores)
+            {
+                if (s.ID != Score.ID)
+                    AddScore(s);
+            }
+
+            hasLoaded = true;
         }
     }
 }
